Refine knowledge index package interface
