+++ .github/scripts/check-knowledge-format/check.ts
@@ -2,7 +2,11 @@ import { readdir, readFile } from "node:fs/promises";
 import { basename, join, relative, resolve, sep } from "node:path";
 import { fileURLToPath } from "node:url";
 
-import { inspectKnowledgeIndex } from "@knowledge-base/knowledge-index";
+import {
+  KnowledgeIndexParseError,
+  parseKnowledgeIndex,
+  type KnowledgeIndexEntry,
+} from "@knowledge-base/knowledge-index";
 import { validateKnowledgeDocument } from "./validate.ts";
 
 export interface KnowledgeFormatDiagnostic {
@@ -36,6 +40,33 @@ async function findMarkdownFiles(directory: string): Promise<string[]> {
 
 function toPortablePath(filePath: string): string {
   return filePath.split(sep).join("/");
+}
+
+function validateIndexCoverage(
+  entries: KnowledgeIndexEntry[],
+  leafFilePaths: string[],
+): string[] {
+  const diagnostics: string[] = [];
+  const leafPathSet = new Set(leafFilePaths);
+  const indexedPathSet = new Set(entries.map((entry) => entry.filePath));
+
+  for (const entry of entries) {
+    if (!leafPathSet.has(entry.filePath)) {
+      diagnostics.push(
+        `The index lists 'knowledge/${entry.filePath}', but that leaf document does not exist.`,
+      );
+    }
+  }
+
+  for (const leafFilePath of leafFilePaths) {
+    if (!indexedPathSet.has(leafFilePath)) {
+      diagnostics.push(
+        `Knowledge leaf 'knowledge/${leafFilePath}' must be listed exactly once in the index.`,
+      );
+    }
+  }
+
+  return diagnostics;
 }
 
 export async function checkKnowledgeDirectory(
@@ -94,10 +125,18 @@ export async function checkKnowledgeDirectory(
     const leafFilePaths = relativeMarkdownFiles.filter(
       (filePath) => filePath !== "index.md" && !filePath.endsWith("/index.md"),
     );
-    const inspection = inspectKnowledgeIndex(indexMarkdown, leafFilePaths);
-
-    for (const message of inspection.diagnostics) {
-      diagnostics.push({ filePath: indexDisplayPath, message });
+    try {
+      const entries = parseKnowledgeIndex(indexMarkdown);
+      for (const message of validateIndexCoverage(entries, leafFilePaths)) {
+        diagnostics.push({ filePath: indexDisplayPath, message });
+      }
+    } catch (error) {
+      if (!(error instanceof KnowledgeIndexParseError)) {
+        throw error;
+      }
+      for (const message of error.diagnostics) {
+        diagnostics.push({ filePath: indexDisplayPath, message });
+      }
     }
   }
