+++ .github/scripts/content-verification/targets.test.ts
@@ -83,20 +83,19 @@ describe("discoverVerificationTargets", () => {
     );
   });
 
-  it("fails when the parsed Knowledge index is incomplete", () => {
+  it("fails when the parsed index and tracked Knowledge differ", () => {
     assert.throws(
       () =>
         discoverVerificationTargets(
           "evergreen-knowledge",
-          [
-            "knowledge/index.md",
-            "knowledge/a.md",
-            "knowledge/b.md",
-            "knowledge/c.md",
-          ],
+          ["knowledge/index.md", "knowledge/a.md", "knowledge/c.md"],
           index,
         ),
-      /invalid index/u,
+      new Error(
+        "Cannot select Knowledge because the index does not match tracked Knowledge:\n" +
+          "The index lists 'knowledge/b.md', but that tracked leaf document does not exist.\n" +
+          "Tracked Knowledge leaf 'knowledge/c.md' must be listed exactly once in the index.",
+      ),
     );
   });
 });
