+++ packages/knowledge-index/types.ts
@@ -1,0 +1,7 @@
+export type KnowledgeType = "time-sensitive" | "evergreen";
+
+export interface KnowledgeIndexEntry {
+  filePath: string;
+  knowledgeType: KnowledgeType;
+  whenToRead: string;
+}
